make national id a required field
