+++ src/form/birth/index.ts
@@ -23,7 +23,8 @@ import {
   getFamilyNameField,
   getFirstNameField,
   getNationality,
-  otherInformantType
+  otherInformantType,
+  getNationalID
 } from '../common/common-required-fields'
 import {
   exactDateOfBirthUnknown,
@@ -31,8 +32,7 @@ import {
   getMaritalStatus,
   registrationEmail,
   registrationPhone,
-  divider,
-  getNationalID
+  divider
 } from '../common/common-optional-fields'
 import {
   attendantAtBirth,
